+++ src/core/moves/admixture/AdmixtureEdgeReplaceCladeResiduals.h
@@ -33,7 +33,7 @@ namespace RevBayesCore {
         // Basic utility functions
         AdmixtureEdgeReplaceCladeResiduals*              clone(void) const;                                                                  //!< Clone object
         void                            swapNode(DagNode *oldN, DagNode *newN);
-        bool                            isActive(int g) const;
+        bool                            isActive(unsigned long g) const;
         
     protected:
         const std::string&              getMoveName(void) const;                                                            //!< Get the name of the move for summary printing
@@ -54,7 +54,7 @@ namespace RevBayesCore {
         std::vector< ContinuousStochasticNode*> branchRates;
         DeterministicNode<std::vector<double> >* residuals;
         //ConstantNode<int>*              delayTimer;
-        int                             activeGen;
+        unsigned long                    activeGen;
         bool                            changed;
         bool                            failed;
         bool                            failedAdd;
@@ -62,8 +62,8 @@ namespace RevBayesCore {
         int                             delay;
         int                             maxEvents;
         int                             numEvents;
-        int                             numNodes;
-        int                             numTips;
+        size_t                             numNodes;
+        size_t                             numTips;
         
         // stored objects to undo proposal
         AdmixtureNode*                  storedAdmixtureParent;
